Todo está funcional y compilando

+++ libft/ft_memset.c
@@ -10,9 +10,11 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_memset(void *str, int value, int n)
+#include <stddef.h>
+
+void	ft_memset(void *str, int value, size_t n)
 {
-	int				i;
+	size_t				i;
 	unsigned char	*ptr;
 
 	ptr = (unsigned char *)str;
